pyproject.toml
[project]
name = "skill-boost-testing"
version = "0.1.0"
description = "Skill boost session about unit testing"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pytest>=8.3.2",
    "mysql-connector-python==9.3.0",
]